pc/requirements.txt
pyserial>=3.5
numpy>=1.24
scipy>=1.10
pandas>=1.5
neurokit2>=0.2
matplotlib>=3.7
cvxopt>=1.3
